+++ src/steinhart_hart.rs
@@ -3,14 +3,14 @@ use lexical_core::Float;
 /// Steinhart-Hart equation parameters
 #[derive(Clone, Debug)]
 pub struct Parameters {
-    pub a: f32,
-    pub b: f32,
-    pub c: f32,
+    pub a: f64,
+    pub b: f64,
+    pub c: f64,
     /// Parallel resistance
     ///
     /// Not truly part of the equation but required to calculate
     /// resistance from voltage.
-    pub parallel_r: f32,
+    pub parallel_r: f64,
 }
 
 impl Parameters {
@@ -19,7 +19,7 @@ impl Parameters {
     /// Result unit: Kelvin
     ///
     /// TODO: verify
-    pub fn get_temperature(&self, voltage: f32) -> f32 {
+    pub fn get_temperature(&self, voltage: f64) -> f64 {
         let r = self.parallel_r * voltage;
         let ln_r = r.abs().ln();
         let inv_temp = self.a +
